Move types directory
