+++ components/DetailPanel.tsx
@@ -2,10 +2,10 @@ import { Fragment, useState, useRef, useEffect, useCallback, MouseEventHandler }
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { ReactFocusEvent } from '@/types/common';
-import { isOneOf } from '@/types/guard';
-import { TodoItem } from '@/types/store/todoSlice';
-import { SettingsPerPage } from '@/types/store/todoSlice';
+import { ReactFocusEvent } from '@/lib/types/common';
+import { isOneOf } from '@/lib/types/guard';
+import { TodoItem } from '@/lib/types/store/todoSlice';
+import { SettingsPerPage } from '@/lib/types/store/todoSlice';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import {
   openDeadlinePicker,
