+++ components/NavigationDrawer.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { isOneOf } from '@/types/guard';
+import { isOneOf } from '@/lib/types/guard';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import { openSidebar, closeSidebar } from '@/lib/store/todoSlice';
 import dayjs from '@/lib/plugins/dayjs';
