+++ components/OrderingCriterion.tsx
@@ -2,8 +2,8 @@ import { useRef, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { OrderingCriterion as OrderingCriterionType, OrderingDirection } from '@/types/common';
-import { isOneOf } from '@/types/guard';
+import { OrderingCriterion as OrderingCriterionType, OrderingDirection } from '@/lib/types/common';
+import { isOneOf } from '@/lib/types/guard';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import {
   IMPORTANCE,
