+++ components/OrderingIndicator.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { OrderingCriterion } from '@/types/common';
-import { isOneOf } from '@/types/guard';
+import { OrderingCriterion } from '@/lib/types/common';
+import { isOneOf } from '@/lib/types/guard';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import {
   IMPORTANCE,
