+++ components/StepInput.tsx
@@ -2,8 +2,8 @@ import { useRef, KeyboardEventHandler, FormEventHandler, FocusEventHandler } fro
 import { useRouter } from 'next/router';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { isOneOf } from '@/types/guard';
-import { SettingsPerPage } from '@/types/store/todoSlice';
+import { isOneOf } from '@/lib/types/guard';
+import { SettingsPerPage } from '@/lib/types/store/todoSlice';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import { createTodoItem, createSubStep } from '@/lib/store/todoSlice';
 import styles from './StepInput.module.scss';
