+++ components/StepList.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
-import { Dict, Nullable } from '@/types/common';
-import { isRegExp, isOneOf } from '@/types/guard';
-import { TodoItemBase, TodoItem, FilteringCondition } from '@/types/store/todoSlice';
+import { Dict, Nullable } from '@/lib/types/common';
+import { isRegExp, isOneOf } from '@/lib/types/guard';
+import { TodoItemBase, TodoItem, FilteringCondition } from '@/lib/types/store/todoSlice';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
 import { openDetailPanel, updateSubStep } from '@/lib/store/todoSlice';
 import styles from './TaskList.module.scss'; // shared
