+++ lib/types/common.ts
