+++ lib/types/guard.ts
@@ -1,4 +1,4 @@
-import { Dict } from '@/types/common';
+import { Dict } from '@/lib/types/common';
 
 export const isDict = (arg: unknown): arg is Dict => (
   arg !== undefined && arg !== null && Object.getPrototypeOf(arg) === Object.prototype
